+++ src/main/java/ladysnake/illuminations/mixin/ComposterBlockMixin.java
@@ -1,0 +1,41 @@
+package ladysnake.illuminations.mixin;
+
+import ladysnake.illuminations.common.IlluminationsItems;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ComposterBlock;
+import net.minecraft.block.InventoryProvider;
+import net.minecraft.entity.ItemEntity;
+import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.Hand;
+import net.minecraft.util.hit.BlockHitResult;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import java.util.concurrent.ThreadLocalRandom;
+
+@Mixin(ComposterBlock.class)
+public abstract class ComposterBlockMixin extends Block implements InventoryProvider {
+    public ComposterBlockMixin(Settings settings) {
+        super(settings);
+    }
+
+    @Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/block/ComposterBlock;emptyFullComposter(Lnet/minecraft/block/BlockState;Lnet/minecraft/world/World;Lnet/minecraft/util/math/BlockPos;)Lnet/minecraft/block/BlockState;"), method = "onUse")
+    public void onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult hit, CallbackInfoReturnable<ActionResult> cir) {
+        if (!world.isClient) {
+            double d = (double)(world.random.nextFloat() * 0.7F) + 0.15000000596046448D;
+            double e = (double)(world.random.nextFloat() * 0.7F) + 0.06000000238418579D + 0.6D;
+            double g = (double)(world.random.nextFloat() * 0.7F) + 0.15000000596046448D;
+            ItemEntity itemEntity = new ItemEntity(world, (double)pos.getX() + d, (double)pos.getY() + e, (double)pos.getZ() + g, new ItemStack(IlluminationsItems.BUGBALL, ThreadLocalRandom.current().nextInt(1, 5)));
+            itemEntity.setToDefaultPickupDelay();
+            world.spawnEntity(itemEntity);
+        }
+    }
+
+}